Added ignored files. Recompiled Audio.lib to remove warnings. Fixed some project configurations to remove warnings. Perform static_cast to remove more warnings.

+++ Player.cpp
@@ -59,7 +59,7 @@ void Player::CollisionCheck(float dt)
         }
         else
         {
-            pos.x = ScreenWidth - theSprite.GetWidth();
+            pos.x = static_cast<float>( ScreenWidth - theSprite.GetWidth() );
         }
     }
 }
